requirements.txt
Flask
flask-restful
requests
requests_cache
cassandra-driver